+++ .coveragerc
@@ -1,0 +1,10 @@
+[run]
+omit =
+       # tests
+       tests/*
+
+
+[report]
+omit =
+       # tests
+       tests/*
