+++ tests/test_py4j_api.py
@@ -16,10 +16,6 @@ class Test_api:
         c = CoreseAPI(java_bridge = "PY4J")
         assert(c.java_bridge == "py4j")
 
-    def test_init_jpype_bridge(self):
-        c = CoreseAPI(java_bridge = "jpype")
-        assert(c.java_bridge == "jpype")
-
     def test_bad_bridge(self):
         with pytest.raises(ValueError):
             CoreseAPI(java_bridge="user_mistake")
@@ -30,9 +26,8 @@ class Test_api:
             c = CoreseAPI(corese_path="/this/file/does/not/exists")
             c.loadCorese() # this will crash
 
+    def test_py4j_version(self):
+        c = CoreseAPI(java_bridge = "PY4J")
 
-    def test_bad_location_jpype(self):
-        with pytest.raises(FileNotFoundError):
-            c = CoreseAPI(corese_path="/this/file/does/not/exists",
-                          java_bridge="jpype")
-            c.loadCorese() # this will crash
+        # version not found since corese is  not loaded
+        assert(c.coreseVersion() is None)
